Not so fast... switch the <missingValueIndicators> namespace back to "study" for backward compatibility. Introduce "folder"-namespace element for exports going forward and prefer that on import.
Move missing value indicator writer and importer from study to core.
