+++ api/schemas/folder.xsd
@@ -81,7 +81,10 @@
                 <xs:element name="queries" minOccurs="0" type="exportDirType" />
                 <xs:element name="views" minOccurs="0" type="exportDirType" />
                 <xs:element name="lists" minOccurs="0" type="exportDirType" />
-                <xs:element name="missingValueIndicators" type="missingValueIndicatorsType" minOccurs="0" />
+                <!-- Starting with 22.3, MVIs are exported to this element in the "folder" namespace -->
+                <xs:element name="missingValueIndicator" type="missingValueIndicatorsType" minOccurs="0" />
+                <!-- Pre-22.3 archives exported MVIs in the "study" namespace -->
+                <xs:element name="missingValueIndicators" type="study:missingValueIndicatorsType" minOccurs="0" />
                 <xs:element name="wikis" minOccurs="0" type="exportDirType"/>
                 <xs:element name="etls" minOccurs="0" type="exportDirType"/>
                 <xs:element name="externalSchemas" minOccurs="0" type="exportDirType"/>
